+++ src/clj/net/willcohen/proj/impl/native.clj
@@ -182,21 +182,14 @@
                  pf (locate-proj-file tmpdir)
                  pd (locate-proj-db tmpdir)
                  _ (locate-grids tmpdir)
-                 os (get-os)
-                 tf (when (= os :darwin) (locate-libtiff-file tmpdir))
                  p (.getCanonicalPath (.getParentFile pf))
-                 t (when tf (.getCanonicalPath (.getParentFile tf)))
                  pl (-> (.getName pf)
                         (.replaceFirst "[.][^.]+$" "")
                         (.replaceFirst "lib" ""))
-                 tl (when tf (-> (.getName tf)
-                                 (.replaceFirst "[.][^.]+$" "")
-                                 (.replaceFirst "lib" "")))
                  s (dt-ffi/library-singleton #'fn-defs)]
              (when (dt-ffi/jna-ffi?)
                (System/setProperty "jna.library.path" (.toString tmpdir)))
-             {:file pf :db pd :libtiff-file tf :path p :libtiff-path t :singleton s :libname pl
-              :libname-libtiff tl})
+             {:file pf :db pd :path p :singleton s :libname pl})
            (catch Exception _ proj))))
 
 ;; If native libs fail to load, check: (clojure.java.io/resource "darwin-aarch64/libproj.dylib")